move TestSQLAction to devtools (#4219)
